backend/requirements.txt
fastapi==0.117.1
sqlmodel==0.0.25
pydantic==2.11.9
uvicorn==0.36.0
psycopg2-binary==2.9.9
python-multipart==0.0.9

passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.5.0